requirements.txt
# LLM Course Dependencies

# Core LLM Libraries
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
llama-index>=0.11.0

# API Clients
openai>=1.50.0
anthropic>=0.35.0
google-generativeai>=0.8.0

# Vector Databases
chromadb>=0.5.0
pinecone-client>=5.0.0
weaviate-client>=4.8.0

# Embeddings
sentence-transformers>=3.1.0

# Document Processing
pypdf>=5.0.0
unstructured>=0.15.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# Web & API
fastapi>=0.115.0
uvicorn>=0.31.0
httpx>=0.27.0
requests>=2.32.0

# Data Processing
pandas>=2.2.0
numpy>=2.0.0

# ML & Fine-tuning
torch>=2.4.0
transformers>=4.45.0
peft>=0.13.0
datasets>=3.0.0
accelerate>=0.34.0
bitsandbytes>=0.44.0

# Evaluation
ragas>=0.1.0
deepeval>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.9.0
tenacity>=9.0.0

# Development
jupyter>=1.1.0
jupyterlab>=4.2.0
ipywidgets>=8.1.0

# Inference
vllm>=0.6.0; sys_platform == 'linux'
